Week06/projects/iot-data-pipeline/requirements.txt
sqlalchemy
paho-mqtt
pandas
plotly
